[Studio] Split large step components for max-lines rule
