+++ assets/studio/js/src/modules/data-importer/components/tabs/steps/mapping-step/mapping-item/index.ts
@@ -1,0 +1,16 @@
+/**
+ * This source file is available under the terms of the
+ * Pimcore Open Core License (POCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (https://www.pimcore.com)
+ *  @license    Pimcore Open Core License (POCL)
+ */
+
+export { MappingItem } from './mapping-item'
+export type { MappingItemProps } from './mapping-item'
+export { MappingDropZone } from './mapping-drop-zone'
+export type { MappingDropZoneProps } from './mapping-drop-zone'
+export { MappingItemWithFilter } from './mapping-item-with-filter'
+export type { MappingItemWithFilterProps } from './mapping-item-with-filter'
